Documentation update
Changelog:
Starting now the UFCP code will have 2 versions: one designed to run on
X-Plane 9.70, and one designed to run on X-Plane 10.20+. The .ino file
will have 2 versions accordingly, a _ver9 for the X-Plane 9 compatible
version and a _ver10 for the X-Plane 10 compatible version.

The User's Manual serves as the official documentation for all current
and planned functionalities. Therefore it should contain all current and
planned functionalities to help with the implementation into the
program.

IMPORTANT: No LCD features are currently added to the source code, their
implementation will only come after the command string formats are all
added and tested.